design: 상세페이지 반응형 디자인 구현
