+++ components/detail/article/contents/comments/index.tsx
@@ -8,9 +8,9 @@ export default function Comment({ commentRef }: any) {
   return (
     <div
       id="comment"
-      className="relative flex w-full flex-col items-start gap-32 self-stretch pt-60"
+      className="relative flex w-full flex-col items-start gap-32 self-stretch pt-60 tablet:gap-24"
     >
-      <span>댓글</span>
+      <span className="text-20 tablet:text-18">댓글</span>
       <div className="flex w-full flex-col items-start gap-40 self-stretch">
         <div className="flex w-full flex-col items-start gap-20 self-stretch">
           {<OthersComment />}
@@ -20,7 +20,7 @@ export default function Comment({ commentRef }: any) {
           <RegisterComment />
         </div>
       </div>
-      <div ref={commentRef} className="absolute top-[750px] h-1 w-full"></div>
+      <div ref={commentRef} className="absolute top-[670px] h-1 w-full"></div>
     </div>
   );
 }
