+++ components/detail/article/contents/comments/register/commentOfcomment/index.tsx
@@ -22,7 +22,7 @@ export default function RegistCommentOfComment() {
       <div className="flex w-full flex-col items-start gap-12 self-stretch">
         <div className="flex items-center justify-between self-stretch py-2">
           <div className="flex items-center gap-12">
-            <div className="relative h-32 w-32 rounded-full">
+            <div className="relative h-32 w-32 rounded-full tablet:h-24 tablet:w-24">
               <Image
                 src={
                   "https://i.namu.wiki/i/5i-kQ5O71eVdYRGVSfquZF5NmnBYYRNcK9bFMq-CD9OI5L-faMaFykGuua7N11FgAuTwiW8vlDrNlK9Yx8TGrA.webp"
@@ -31,7 +31,7 @@ export default function RegistCommentOfComment() {
                 fill
               />
             </div>
-            <span className="text-18 leading-[27px] tracking-[-0.6px] text-text-01">
+            <span className="text-18 leading-[27px] tracking-[-0.6px] text-text-01 tablet:text-16 tablet:leading-[20.8px]">
               {"내 닉네임"}
             </span>
           </div>
