+++ components/detail/article/contents/comments/secret/commentOfcomment/index.tsx
@@ -8,7 +8,7 @@ export default function SecretCommentOfComment() {
           <Image src="/icons/frame.svg" alt="대댓글 이미지" fill />
         </div>
         <div className="flex items-start gap-8 py-12">
-          <span className="text-16 font-normal not-italic leading-[20.8px] tracking-[-0.6px] text-text-02">
+          <span className="text-16 font-normal not-italic leading-[20.8px] tracking-[-0.6px] text-text-02 mobile:text-14 mobile:leading-[21px]">
             {"비밀 댓글입니다."}
           </span>
         </div>
