+++ components/detail/article/contents/comments/secret/index.tsx
@@ -4,7 +4,7 @@ export default function SecretCommment() {
   return (
     <>
       <div className="flex items-start gap-8 self-stretch py-12">
-        <span className="overflow-hidden text-ellipsis text-16 font-normal leading-6 tracking-[-0.6px] text-text-02">
+        <span className="overflow-hidden text-ellipsis text-16 font-normal leading-6 tracking-[-0.6px] text-text-02 tablet:text-14 tablet:leading-[21px]">
           비밀 댓글입니다.
         </span>
       </div>
