+++ components/detail/article/contents/destination/index.tsx
@@ -38,25 +38,25 @@ export default function Destination({ destinationRef }: DestinationProps) {
   return (
     <div
       id="destination"
-      className="relative flex w-full flex-col items-start gap-32 self-stretch pt-60"
+      className="relative flex w-full flex-col items-start gap-32 self-stretch pt-60 tablet:gap-24"
     >
-      <span>여행지</span>
+      <span className="text-20 tablet:text-18">여행지</span>
       <div className="flex flex-col items-start gap-16 self-stretch">
         <div className="flex items-center gap-12">
-          <div className="relative h-24 w-24">
+          <div className="relative h-24 w-24 tablet:h-20 tablet:w-20">
             <Image src="/icons/location_red.svg" alt="장소 표시 이미지" fill />
           </div>
-          <span className="text-16 font-bold leading-[20.8px]">
+          <span className="text-16 font-bold leading-[20.8px] tablet:text-14">
             {destination}
           </span>
         </div>
-        <div className="tablet:h-376 h-[500px] w-[892px] rounded-12 tablet:w-[672px] mobile:h-[152.5px] mobile:w-272">
+        <div className="tablet:h-376 h-[500px] w-[892px] rounded-12 tablet:h-[376px] tablet:w-[672px] mobile:h-[152.5px] mobile:w-272">
           <GoogleMap location={location} />
         </div>
       </div>
       <div
         ref={destinationRef}
-        className="absolute top-[750px] h-1 w-full"
+        className="absolute top-[670px] h-1 w-full"
       ></div>
     </div>
   );
