+++ components/detail/article/contents/sort/index.tsx
@@ -8,12 +8,12 @@ export default function DetailSort({
   const [selectedId, setSelectedId] = useState("information");
 
   useEffect(() => {
-    if (recruitInView) {
-      setSelectedId("information");
+    if (commentInView) {
+      setSelectedId("comment");
     } else if (destinationInView) {
       setSelectedId("destination");
-    } else if (commentInView) {
-      setSelectedId("comment");
+    } else if (recruitInView) {
+      setSelectedId("information");
     }
   }, [recruitInView, destinationInView, commentInView]);
 
@@ -32,9 +32,9 @@ export default function DetailSort({
   });
 
   return (
-    <div className="sticky top-0 z-10 flex w-full items-start gap-4 bg-sky-50 text-20 font-bold text-sky-600">
+    <div className="sticky top-0 z-10 flex w-full items-start gap-4 bg-sky-50 text-20 font-bold text-sky-600 tablet:text-18 mobile:text-16">
       <a
-        className="flex h-60 w-1/3 justify-around rounded-tl-32 rounded-tr-32 border-none"
+        className="flex h-60 w-1/3 justify-around rounded-tl-24 rounded-tr-24 border-none tablet:h-56 mobile:h-44"
         href="#information"
         onClick={handleClick("information")}
         style={getLinkStyle("information")}
@@ -42,7 +42,7 @@ export default function DetailSort({
         <span className="flex items-center">모집정보</span>
       </a>
       <a
-        className="flex h-60 w-1/3 justify-around rounded-tl-32 rounded-tr-32 border-none"
+        className="flex h-60 w-1/3 justify-around rounded-tl-24 rounded-tr-24 border-none tablet:h-56 mobile:h-44"
         href="#destination"
         onClick={handleClick("destination")}
         style={getLinkStyle("destination")}
@@ -50,7 +50,7 @@ export default function DetailSort({
         <span className="flex items-center">여행지</span>
       </a>
       <a
-        className="flex h-60 w-1/3 justify-around rounded-tl-32 rounded-tr-32 border-none"
+        className="flex h-60 w-1/3 justify-around rounded-tl-24 rounded-tr-24 border-none tablet:h-56 mobile:h-44"
         href="#comment"
         onClick={handleClick("comment")}
         style={getLinkStyle("comment")}
