+++ components/detail/article/index.tsx
@@ -3,7 +3,7 @@ import Title from "@/components/detail/article/title/index";
 
 export default function Article() {
   return (
-    <div className="self flex w-full flex-col items-center gap-24 self-stretch pb-80">
+    <div className="self mx-auto flex w-full max-w-[956px] flex-col items-center gap-24 self-stretch pb-80 tablet:max-w-[720px] mobile:max-w-[312px]">
       <Title />
       <Content />
     </div>
